reuse array conciler for all types, some cleanup
